+++ MOP/src/GameObjects/Items/Items.cs
@@ -200,7 +200,7 @@ namespace MOP
             itemsObject.GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
 
             // Fucking wheels.
-            GameObject[] wheels = Object.FindObjectsOfType<GameObject>().Where(gm => gm.name == "wheel_regula" && gm.activeSelf).ToArray();
+            GameObject[] wheels = Object.FindObjectsOfType<GameObject>().Where(gm => gm.name.EqualsAny("wheel_regula", "wheel_offset") && gm.activeSelf).ToArray();
             foreach (GameObject wheel in wheels)
                 wheel.AddComponent<ItemHook>();
         }
@@ -215,7 +215,7 @@ namespace MOP
 
             for (int i = 0; i < beerBottles.Length; i++)
             {
-                GameObject.Destroy(beerBottles[i]);
+                Object.Destroy(beerBottles[i]);
             }
         }