+++ MOP/src/GameObjects/Vehicles/Managers/Satsuma/SatsumaSeatsManager.cs
@@ -21,6 +21,10 @@ namespace MOP
 {
     class SatsumaSeatsManager : MonoBehaviour
     {
+        // So this is stupid fix.
+        // Sets the weight of the seats rigidbody to 1, if seat is not bolted.
+        // And sets it to their default value, if the screw tightness level is more than 7.
+
         readonly Rigidbody rb;
         readonly FsmFloat tightness;
         readonly float defaultMass;