+++ MOP/src/RuleFiles/RuleFilesLoader.cs
@@ -22,6 +22,7 @@ using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Sockets;
+using System.Text;
 using UnityEngine;
 
 namespace MOP